+++ src/main/java/data_access/IStockDataAccess.java
@@ -4,6 +4,7 @@ import java.util.Map;
 
 public interface IStockDataAccess {
 
-    public double getStockPrice(String ticker);
-    public Map<String, Double> getStockPrices();
+    double getStockPrice(String ticker);
+
+    Map<String, Double> getStockPrices();
 }