+++ docs/CHANGELOG.md
@@ -1,12 +1,16 @@
 # Project Changelog
 
 ## Unreleased
+
 ### New Features:
+
 ### Bug Fixes:
+
 ### Internal Changes:
 
 - Created documentation for GitHub workflow and pull request template.
 - Added maven workspace and configuration.
-- Added Use Cases 4 and 5
-- Implement entities for stock and stock market
+- Added Use Cases 4 and 5.
+- Implemented entities for stock and stock market.
+- Implemented entities for user, portfolio, and transaction.
 - Add synchronization checks and null checks in StockMarket